Exchange Application settings for project settings only
